backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6